git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@653 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ test/metagen/include/metagen_operator_other.h
@@ -9,7 +9,7 @@ class MetagenOperatorOther
 {
 public:
 	MetagenOperatorOther() : value(0) {}
-	explicit MetagenOperatorOther(int value) : value(value), f(value) {}
+	explicit MetagenOperatorOther(int value) : value(value), f((float)value) {}
 
 	MetagenOperatorOther & operator = (const MetagenOperatorOther & other) { value = other.value; return *this; }
